feat: html method support new structure
