--- packages/pwc/src/elements/updateView.ts
@@ -1,32 +1,0 @@
-import type { PWCElementTemplate, TemplateNodeValue } from '../type';
-import { shallowEqual } from '../utils';
-import { ReactiveNode, TemplateNode } from './reactiveNode';
-
-export function updateView(
-  oldElementTemplate: PWCElementTemplate,
-  newElementTemplate: PWCElementTemplate,
-  reactiveNodes: ReactiveNode[],
-) {
-  const {
-    templateString: oldTemplateString,
-    templateData: oldTemplateData,
-  } = oldElementTemplate;
-  const {
-    templateString,
-    templateData,
-  } = newElementTemplate;
-  // While template strings is constant with prev ones,
-  // it should just update node values and attributes
-  if (oldTemplateString === templateString) {
-    for (let index = 0; index < oldTemplateData.length; index++) {
-      const reactiveNode = reactiveNodes[index];
-      // Avoid html fragment effect
-      if (reactiveNode instanceof TemplateNode) {
-        // TODO more diff
-        reactiveNode.commitValue([oldTemplateData[index], templateData[index]] as TemplateNodeValue);
-      } else if (!shallowEqual(oldTemplateData[index], templateData[index])) {
-        reactiveNode.commitValue(templateData[index]);
-      }
-    }
-  }
-}
