+++ packages/pwc/src/elements/validateElementTemplate.ts
@@ -3,6 +3,7 @@ import { ElementTemplate } from '../type';
 
 export function validateElementTemplate(elementTemplate: ElementTemplate) {
   if (typeof elementTemplate === 'string') return;
+  console.log(elementTemplate)
 
   if (elementTemplate.template !== true) {
     throwError('Invalid template, it should return type field.');
